# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Viv Scanner - Firmware Security Analysis Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /firmware:
    get:
      operationId: listFirmware
      tags: [firmware]
      summary: List all firmware uploads
      responses:
        "200":
          description: List of firmware
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Firmware"

    post:
      operationId: uploadFirmware
      tags: [firmware]
      summary: Upload firmware file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirmwareInput"
      responses:
        "201":
          description: Firmware uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Firmware"
  /firmware/{id}:
    get:
      operationId: getFirmware
      tags: [firmware]
      summary: Get firmware details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Firmware details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Firmware"
        "404":
          description: Not found

    delete:
      operationId: deleteFirmware
      tags: [firmware]
      summary: Delete firmware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/pdf/{firmwareId}:
    get:
      operationId: getPdfReport
      tags: [reports]
      summary: Get PDF report for a firmware scan
      parameters:
        - name: firmwareId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF report info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportMeta"
  /reports/ai-summary/{firmwareId}:
    get:
      operationId: getAiSummary
      tags: [reports]
      summary: Get AI-generated summary and risk analysis
      parameters:
        - name: firmwareId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReport"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, analyst, viewer]
        createdAt:
          type: string
    UserRegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    UserLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Firmware:
      type: object
      required: [id, name, uploadedAt, architecture, hashValue, status, fileSize]
      properties:
        id:
          type: integer
        name:
          type: string
        uploadedAt:
          type: string
        architecture:
          type: string
          enum: [ARM, MIPS, x86, x86_64, UNKNOWN]
        hashValue:
          type: string
        status:
          type: string
          enum: [pending, scanning, completed, failed]
        fileSize:
          type: integer
        vendor:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
    FirmwareInput:
      type: object
      required: [name, hashValue, fileSize]
      properties:
        name:
          type: string
        hashValue:
          type: string
        fileSize:
          type: integer
        architecture:
          type: string
        vendor:
          type: string
        version:
          type: string
...
    ReportMeta:
      type: object
      required: [firmwareId, generatedAt, downloadUrl, fileSize]
      properties:
        firmwareId:
          type: integer
        generatedAt:
          type: string
        downloadUrl:
          type: string
        fileSize:
          type: integer
        scanId:
          type: ["integer", "null"]
    AiReport:
      type: object
      required: [firmwareId, summary, riskLevel, keyFindings, recommendations, generatedAt]
      properties:
        firmwareId:
          type: integer
        summary:
          type: string
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        keyFindings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        generatedAt:
          type: string
        exploitProbability:
          type: ["number", "null"]
...